+++ CHANGELOG.md
@@ -9,6 +9,13 @@
   pattern synonym, for backwards compatibility.  It will be removed
   in a future release.
 
+- Implemented **reportable fringe benefits**. .  Use the
+  `fringeBenefitsEmployerNotExempt` or
+  `fringeBenefitsEmployerExempt` constructors, along with the
+  `reportableFringeBenefits` lens to add it to the payment summary.
+  The amount is properly included in the surcharge and student loan
+  repayment income calculations.
+
 - New module `Data.Tax.ATO.Depreciation` provides types and
   functions for calculating decline in value (for deductions).
   Use the `DepreciatingAsset` data type to define depreciating